# swagger.yml
openapi: 3.0.2
info:
  title: OnlineAppointment2
  version: 1.0.0
paths:
  /users:
    summary: Path used to manage the list of users.
    description: >-
      The REST endpoint/path used to list and create zero or more `User` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `User` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      tags:
        - /users
      responses:
        '200':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                'Invalid email: {reason}':
                  value: 'Invalid email: {reason}'
                Email and password are required:
                  value: Email and password are required
                Password must be at least 8 characters long:
                  value: Password must be at least 8 characters long
                Email must be in a valid format:
                  value: Email must be in a valid format
          description: Client side error
        '409':
          content:
            text/html:
              examples:
                Email already exists:
                  value: Email already exists
          description: The email address provided already exists in the system.
      operationId: createUser
      summary: Create a User
      description: Creates a new instance of a `User`.
  /sessions:
    summary: Path used to manage the list of session
    description: >-
      The REST endpoint/path used to list and create zero or more Session entities. The path contains
      POST operation to perform the list and create task, respectively.
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
        required: true
      tags:
        - /sessions
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: Successful response
        '400':
          content:
            text/xml:
              examples:
                Bad request:
                  value: Email and password a required
          description: Bad request
        '401':
          content:
            text/xml:
              examples:
                Invalid password:
                  value: Invalid password
          description: Invalid password
        '404':
          content:
            text/xml:
              examples:
                404 Not Found:
                  value: User not found
          description: |
            User not found
      summary: List all sessions
      description: Get a list of all session entities
    delete:
      tags:
        - /sessions
      responses:
        '204':
          description: Delete successful response
        '400':
          content:
            text/xml:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Invalid authorization header
        '401':
          content:
            text/xml:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Missing authorization header
      security:
        -
          Bearer: []
      summary: Destroy a session
      description: 'Destroys a session, signs the user out. Returns session id.'
  /appointments:
    summary: Array of Appointment entities
    description: Gets a list of all `Appointment` entities.
    get:
      tags:
        - /appointments
      responses:
        '200':
          content:
            text/xml:
              examples:
                Successful response - returns an array of `Appointment` entities.:
                  value: Successful response - returns an array of `Appointment` entities.
          description: Successful response - returns an array of `Appointment` entities.
        '400':
          content:
            text/xml:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad request
        '401':
          content:
            text/xml:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Missing authorization header
      summary: Gets a list of all `Appointment` entities
      description: Gets a list of all `Appointment` entities.
    post:
      tags:
        - /appointments
      responses:
        '201':
          content:
            text/xml:
              examples:
                Successful response:
                  value: Successful response
          description: Successful response
        '400':
          content:
            text/xml:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Invalid authorization header
        '401':
          content:
            text/xml:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Missing authorization header
      security:
        -
          Bearer: []
      summary: Creates a new instance of a `Appointment`
      description: Creates a new instance of a `Appointment`
components:
  schemas:
    User:
      title: Root Type for User
      description: ''
      type: object
      properties:
        id:
          format: int32
          description: A field that represents the unique identifier of the user.
          type: integer
          example: '12345'
        email:
          description: A field that represents the email address of the user.
          type: string
          example: john.doe@example.com
        password:
          description: A field that represents the password of the user.
          type: string
          example: MySecurePassword123
      example:
        id: 3
        email: email@example.com
        password: pass123
    Session:
      title: Root Type for Session
      description: ''
      type: object
      properties:
        id:
          type: string
        userId:
          format: int32
          type: integer
      example:
        id: 330d67de-432b-48ce-a617-0bd6774ebcee
        userId: 1
    Credentials:
      title: Root Type for Credentials
      description: ''
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: admin
        password: KollneKollne
    Appointment:
      title: Root Type for Appointment
      description: ''
      type: object
      properties:
        content:
          type: string
        id:
          format: int32
          type: integer
        title:
          type: string
        userId:
          format: int32
          type: integer
      example:
        content: This is the content of appointment 1
        id: 1
        title: Appointment 1
        userId: 1
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
tags:
  -
    name: /users
    description: ''
  -
    name: /sessions
    description: ''
  -
    name: /appointments
    description: ''